Test SQL database
In this version, I give a try to sql databse.
Finally, I will use NoSql (Mongo) because, for a such simple application, sql requests are too numerous and where the database will be larger, the final user will feel the latency.
NoSql databases can evolve better too.

This version can not be compiled because of errors. I will probably return on it, when I will have more time.